+++ ddm-types/src/db.rs
@@ -1,0 +1,90 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+use std::net::Ipv6Addr;
+
+use mg_common::net::TunnelOrigin;
+use schemars::{JsonSchema, JsonSchema_repr};
+use serde::{Deserialize, Serialize};
+use serde_repr::{Deserialize_repr, Serialize_repr};
+
+#[derive(
+    Debug, Copy, Clone, PartialEq, Eq, Deserialize, Serialize, JsonSchema,
+)]
+pub enum PeerStatus {
+    NoContact,
+    Active,
+    Expired,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema, PartialEq)]
+pub struct PeerInfo {
+    pub status: PeerStatus,
+    pub addr: Ipv6Addr,
+    pub host: String,
+    pub kind: RouterKind,
+}
+
+#[derive(
+    Debug,
+    PartialEq,
+    Eq,
+    Clone,
+    Copy,
+    Serialize_repr,
+    Deserialize_repr,
+    JsonSchema_repr,
+)]
+#[repr(u8)]
+pub enum RouterKind {
+    Server,
+    Transit,
+}
+
+impl std::fmt::Display for RouterKind {
+    fn fmt(
+        &self,
+        f: &mut std::fmt::Formatter<'_>,
+    ) -> Result<(), std::fmt::Error> {
+        match self {
+            Self::Server => write!(f, "server"),
+            Self::Transit => write!(f, "transit"),
+        }
+    }
+}
+
+impl std::str::FromStr for RouterKind {
+    type Err = &'static str;
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "server" => Ok(Self::Server),
+            "transit" => Ok(Self::Transit),
+            _ => Err(r#"Router kind must be "server" or "transit""#),
+        }
+    }
+}
+
+#[derive(
+    Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize, JsonSchema,
+)]
+pub struct TunnelRoute {
+    pub origin: TunnelOrigin,
+
+    // The nexthop is only used to associate the route with a peer allowing us
+    // to remove the route if the peer expires. It does not influence what goes
+    // into the underlaying underlay routing platform. Tunnel routes only
+    // influence the state of the underlying encapsulation service.
+    pub nexthop: Ipv6Addr,
+}
+
+impl From<TunnelRoute> for TunnelOrigin {
+    fn from(x: TunnelRoute) -> Self {
+        Self {
+            overlay_prefix: x.origin.overlay_prefix,
+            boundary_addr: x.origin.boundary_addr,
+            vni: x.origin.vni,
+            metric: x.origin.metric,
+        }
+    }
+}
