+++ ddm/Cargo.toml
@@ -16,7 +16,6 @@ thiserror.workspace = true
 dropshot.workspace = true
 schemars.workspace = true
 tokio.workspace = true
-serde_repr.workspace = true
 anyhow.workspace = true
 hyper.workspace = true
 hyper-util.workspace = true
@@ -36,3 +35,5 @@ oxnet.workspace = true
 uuid.workspace = true
 reqwest.workspace = true
 semver.workspace = true
+ddm-api.workspace = true
+ddm-types.workspace = true
