+++ ddm/src/admin.rs
@@ -2,10 +2,11 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-use crate::db::{Db, PeerInfo, TunnelRoute};
-use crate::exchange::PathVector;
+use crate::db::Db;
 use crate::sm::{AdminEvent, Event, PrefixSet, SmContext};
-use dropshot::ApiDescription;
+use ddm_api::*;
+use ddm_types::db::{PeerInfo, TunnelRoute};
+use ddm_types::exchange::PathVector;
 use dropshot::ConfigDropshot;
 use dropshot::ConfigLogging;
 use dropshot::ConfigLoggingLevel;
@@ -16,22 +17,21 @@ use dropshot::HttpServerStarter;
 use dropshot::Path;
 use dropshot::RequestContext;
 use dropshot::TypedBody;
-use dropshot::{endpoint, ApiDescriptionRegisterError};
+use dropshot::{ApiDescription, ApiDescriptionBuildErrors};
 use mg_common::lock;
 use mg_common::net::TunnelOrigin;
 use oxnet::Ipv6Net;
-use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
 use slog::{error, info, warn, Logger};
-use std::collections::{BTreeMap, HashMap, HashSet};
-use std::net::{IpAddr, Ipv6Addr, SocketAddr, SocketAddrV4, SocketAddrV6};
+use std::collections::{HashMap, HashSet};
+use std::net::{IpAddr, SocketAddr, SocketAddrV4, SocketAddrV6};
 use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::mpsc::Sender;
 use std::sync::Arc;
 use std::sync::Mutex;
 use tokio::spawn;
 use tokio::task::JoinHandle;
-use uuid::Uuid;
+
+pub const DDM_STATS_PORT: u16 = 8001;
 
 #[derive(Default)]
 pub struct RouterStats {
@@ -91,334 +91,299 @@ pub fn handler(
     Ok(())
 }
 
-#[endpoint { method = GET, path = "/peers" }]
-async fn get_peers(
-    ctx: RequestContext<Arc<Mutex<HandlerContext>>>,
-) -> Result<HttpResponseOk<HashMap<u32, PeerInfo>>, HttpError> {
-    let ctx = lock!(ctx.context());
-    Ok(HttpResponseOk(ctx.db.peers()))
-}
+pub enum DdmAdminApiImpl {}
 
-#[derive(Debug, Clone, Deserialize, Serialize, JsonSchema)]
-struct ExpirePathParams {
-    addr: Ipv6Addr,
-}
+impl DdmAdminApi for DdmAdminApiImpl {
+    type Context = Arc<Mutex<HandlerContext>>;
 
-#[endpoint { method = DELETE, path = "/peers/{addr}" }]
-async fn expire_peer(
-    ctx: RequestContext<Arc<Mutex<HandlerContext>>>,
-    params: Path<ExpirePathParams>,
-) -> Result<HttpResponseUpdatedNoContent, HttpError> {
-    let addr = params.into_inner().addr;
-    let ctx = lock!(ctx.context());
+    async fn get_peers(
+        ctx: RequestContext<Self::Context>,
+    ) -> Result<HttpResponseOk<HashMap<u32, PeerInfo>>, HttpError> {
+        let ctx = lock!(ctx.context());
+        Ok(HttpResponseOk(ctx.db.peers()))
+    }
 
-    for e in &ctx.event_channels {
-        e.send(Event::Admin(AdminEvent::Expire(addr)))
+    async fn expire_peer(
+        ctx: RequestContext<Self::Context>,
+        params: Path<ExpirePathParams>,
+    ) -> Result<HttpResponseUpdatedNoContent, HttpError> {
+        let addr = params.into_inner().addr;
+        let ctx = lock!(ctx.context());
+
+        for e in &ctx.event_channels {
+            e.send(Event::Admin(AdminEvent::Expire(addr)))
+                .map_err(|e| {
+                    HttpError::for_internal_error(format!(
+                        "admin event send: {e}"
+                    ))
+                })?;
+        }
+
+        Ok(HttpResponseUpdatedNoContent())
+    }
+
+    async fn get_originated(
+        ctx: RequestContext<Self::Context>,
+    ) -> Result<HttpResponseOk<HashSet<Ipv6Net>>, HttpError> {
+        let ctx = lock!(ctx.context());
+        let originated = ctx
+            .db
+            .originated()
+            .map_err(|e| HttpError::for_internal_error(e.to_string()))?;
+        Ok(HttpResponseOk(originated))
+    }
+
+    async fn get_originated_tunnel_endpoints(
+        ctx: RequestContext<Self::Context>,
+    ) -> Result<HttpResponseOk<HashSet<TunnelOrigin>>, HttpError> {
+        let ctx = lock!(ctx.context());
+        let originated = ctx
+            .db
+            .originated_tunnel()
+            .map_err(|e| HttpError::for_internal_error(e.to_string()))?;
+        Ok(HttpResponseOk(originated))
+    }
+
+    async fn get_prefixes(
+        ctx: RequestContext<Self::Context>,
+    ) -> Result<HttpResponseOk<PrefixMap>, HttpError> {
+        let ctx = lock!(ctx.context());
+        let imported = ctx.db.imported();
+
+        let mut result = PrefixMap::default();
+
+        for route in imported {
+            if let Some(entry) = result.get_mut(&route.nexthop) {
+                entry.insert(PathVector {
+                    destination: route.destination,
+                    path: route.path,
+                });
+            } else {
+                let mut s = HashSet::new();
+                s.insert(PathVector {
+                    destination: route.destination,
+                    path: route.path,
+                });
+                result.insert(route.nexthop, s);
+            }
+        }
+
+        Ok(HttpResponseOk(result))
+    }
+
+    async fn get_tunnel_endpoints(
+        ctx: RequestContext<Self::Context>,
+    ) -> Result<HttpResponseOk<HashSet<TunnelRoute>>, HttpError> {
+        let ctx = lock!(ctx.context());
+        let imported = ctx.db.imported_tunnel();
+        Ok(HttpResponseOk(imported))
+    }
+
+    async fn advertise_prefixes(
+        ctx: RequestContext<Self::Context>,
+        request: TypedBody<HashSet<Ipv6Net>>,
+    ) -> Result<HttpResponseUpdatedNoContent, HttpError> {
+        let ctx = lock!(ctx.context());
+        let prefixes = request.into_inner();
+        ctx.db
+            .originate(&prefixes)
+            .map_err(|e| HttpError::for_internal_error(e.to_string()))?;
+
+        for e in &ctx.event_channels {
+            e.send(Event::Admin(AdminEvent::Announce(PrefixSet::Underlay(
+                prefixes.clone(),
+            ))))
             .map_err(|e| {
                 HttpError::for_internal_error(format!("admin event send: {e}"))
             })?;
-    }
-
-    Ok(HttpResponseUpdatedNoContent())
-}
-
-type PrefixMap = BTreeMap<Ipv6Addr, HashSet<PathVector>>;
-
-#[endpoint { method = GET, path = "/originated" }]
-async fn get_originated(
-    ctx: RequestContext<Arc<Mutex<HandlerContext>>>,
-) -> Result<HttpResponseOk<HashSet<Ipv6Net>>, HttpError> {
-    let ctx = lock!(ctx.context());
-    let originated = ctx
-        .db
-        .originated()
-        .map_err(|e| HttpError::for_internal_error(e.to_string()))?;
-    Ok(HttpResponseOk(originated))
-}
-
-#[endpoint { method = GET, path = "/originated_tunnel_endpoints" }]
-async fn get_originated_tunnel_endpoints(
-    ctx: RequestContext<Arc<Mutex<HandlerContext>>>,
-) -> Result<HttpResponseOk<HashSet<TunnelOrigin>>, HttpError> {
-    let ctx = lock!(ctx.context());
-    let originated = ctx
-        .db
-        .originated_tunnel()
-        .map_err(|e| HttpError::for_internal_error(e.to_string()))?;
-    Ok(HttpResponseOk(originated))
-}
-
-#[endpoint { method = GET, path = "/prefixes" }]
-async fn get_prefixes(
-    ctx: RequestContext<Arc<Mutex<HandlerContext>>>,
-) -> Result<HttpResponseOk<PrefixMap>, HttpError> {
-    let ctx = lock!(ctx.context());
-    let imported = ctx.db.imported();
-
-    let mut result = PrefixMap::default();
-
-    for route in imported {
-        if let Some(entry) = result.get_mut(&route.nexthop) {
-            entry.insert(PathVector {
-                destination: route.destination,
-                path: route.path,
-            });
-        } else {
-            let mut s = HashSet::new();
-            s.insert(PathVector {
-                destination: route.destination,
-                path: route.path,
-            });
-            result.insert(route.nexthop, s);
         }
-    }
 
-    Ok(HttpResponseOk(result))
-}
-
-#[endpoint { method = GET, path = "/tunnel_endpoints" }]
-async fn get_tunnel_endpoints(
-    ctx: RequestContext<Arc<Mutex<HandlerContext>>>,
-) -> Result<HttpResponseOk<HashSet<TunnelRoute>>, HttpError> {
-    let ctx = lock!(ctx.context());
-    let imported = ctx.db.imported_tunnel();
-    Ok(HttpResponseOk(imported))
-}
-
-#[endpoint { method = PUT, path = "/prefix" }]
-async fn advertise_prefixes(
-    ctx: RequestContext<Arc<Mutex<HandlerContext>>>,
-    request: TypedBody<HashSet<Ipv6Net>>,
-) -> Result<HttpResponseUpdatedNoContent, HttpError> {
-    let ctx = lock!(ctx.context());
-    let prefixes = request.into_inner();
-    ctx.db
-        .originate(&prefixes)
-        .map_err(|e| HttpError::for_internal_error(e.to_string()))?;
-
-    for e in &ctx.event_channels {
-        e.send(Event::Admin(AdminEvent::Announce(PrefixSet::Underlay(
-            prefixes.clone(),
-        ))))
-        .map_err(|e| {
-            HttpError::for_internal_error(format!("admin event send: {e}"))
-        })?;
-    }
-
-    match ctx.db.originated_count() {
-        Ok(count) => ctx
-            .stats
-            .originated_underlay_prefixes
-            .store(count as u64, Ordering::Relaxed),
-        Err(e) => {
-            error!(
-                ctx.log,
-                "failed to update originated underlay prefixes stat: {e}"
-            )
+        match ctx.db.originated_count() {
+            Ok(count) => ctx
+                .stats
+                .originated_underlay_prefixes
+                .store(count as u64, Ordering::Relaxed),
+            Err(e) => {
+                error!(
+                    ctx.log,
+                    "failed to update originated underlay prefixes stat: {e}"
+                )
+            }
         }
+
+        Ok(HttpResponseUpdatedNoContent())
     }
 
-    Ok(HttpResponseUpdatedNoContent())
-}
+    async fn advertise_tunnel_endpoints(
+        ctx: RequestContext<Self::Context>,
+        request: TypedBody<HashSet<TunnelOrigin>>,
+    ) -> Result<HttpResponseUpdatedNoContent, HttpError> {
+        let ctx = lock!(ctx.context());
+        let endpoints = request.into_inner();
+        slog::info!(ctx.log, "advertise tunnel: {:#?}", endpoints);
+        ctx.db
+            .originate_tunnel(&endpoints)
+            .map_err(|e| HttpError::for_internal_error(e.to_string()))?;
 
-#[endpoint { method = PUT, path = "/tunnel_endpoint" }]
-async fn advertise_tunnel_endpoints(
-    ctx: RequestContext<Arc<Mutex<HandlerContext>>>,
-    request: TypedBody<HashSet<TunnelOrigin>>,
-) -> Result<HttpResponseUpdatedNoContent, HttpError> {
-    let ctx = lock!(ctx.context());
-    let endpoints = request.into_inner();
-    slog::info!(ctx.log, "advertise tunnel: {:#?}", endpoints);
-    ctx.db
-        .originate_tunnel(&endpoints)
-        .map_err(|e| HttpError::for_internal_error(e.to_string()))?;
-
-    for e in &ctx.event_channels {
-        e.send(Event::Admin(AdminEvent::Announce(PrefixSet::Tunnel(
-            endpoints.clone(),
-        ))))
-        .map_err(|e| {
-            HttpError::for_internal_error(format!("admin event send: {e}"))
-        })?;
-    }
-
-    match ctx.db.originated_tunnel_count() {
-        Ok(count) => ctx
-            .stats
-            .originated_tunnel_endpoints
-            .store(count as u64, Ordering::Relaxed),
-        Err(e) => {
-            error!(
-                ctx.log,
-                "failed to update originated tunnel endpoints stat: {e}"
-            )
-        }
-    }
-    Ok(HttpResponseUpdatedNoContent())
-}
-
-#[endpoint { method = DELETE, path = "/prefix" }]
-async fn withdraw_prefixes(
-    ctx: RequestContext<Arc<Mutex<HandlerContext>>>,
-    request: TypedBody<HashSet<Ipv6Net>>,
-) -> Result<HttpResponseUpdatedNoContent, HttpError> {
-    let ctx = lock!(ctx.context());
-    let prefixes = request.into_inner();
-    ctx.db
-        .withdraw(&prefixes)
-        .map_err(|e| HttpError::for_internal_error(e.to_string()))?;
-
-    for e in &ctx.event_channels {
-        e.send(Event::Admin(AdminEvent::Withdraw(PrefixSet::Underlay(
-            prefixes.clone(),
-        ))))
-        .map_err(|e| {
-            HttpError::for_internal_error(format!("admin event send: {e}"))
-        })?;
-    }
-
-    match ctx.db.originated_count() {
-        Ok(count) => ctx
-            .stats
-            .originated_underlay_prefixes
-            .store(count as u64, Ordering::Relaxed),
-        Err(e) => {
-            error!(
-                ctx.log,
-                "failed to update originated underlay prefixes stat: {e}"
-            )
-        }
-    }
-
-    Ok(HttpResponseUpdatedNoContent())
-}
-
-#[endpoint { method = DELETE, path = "/tunnel_endpoint" }]
-async fn withdraw_tunnel_endpoints(
-    ctx: RequestContext<Arc<Mutex<HandlerContext>>>,
-    request: TypedBody<HashSet<TunnelOrigin>>,
-) -> Result<HttpResponseUpdatedNoContent, HttpError> {
-    let ctx = lock!(ctx.context());
-    let endpoints = request.into_inner();
-    slog::info!(ctx.log, "withdraw tunnel: {:#?}", endpoints);
-    ctx.db
-        .withdraw_tunnel(&endpoints)
-        .map_err(|e| HttpError::for_internal_error(e.to_string()))?;
-
-    for e in &ctx.event_channels {
-        e.send(Event::Admin(AdminEvent::Withdraw(PrefixSet::Tunnel(
-            endpoints.clone(),
-        ))))
-        .map_err(|e| {
-            HttpError::for_internal_error(format!("admin event send: {e}"))
-        })?;
-    }
-
-    match ctx.db.originated_tunnel_count() {
-        Ok(count) => ctx
-            .stats
-            .originated_tunnel_endpoints
-            .store(count as u64, Ordering::Relaxed),
-        Err(e) => {
-            error!(
-                ctx.log,
-                "failed to update originated tunel endpoints stat: {e}"
-            )
-        }
-    }
-
-    Ok(HttpResponseUpdatedNoContent())
-}
-
-#[endpoint { method = PUT, path = "/sync" }]
-async fn sync(
-    ctx: RequestContext<Arc<Mutex<HandlerContext>>>,
-) -> Result<HttpResponseUpdatedNoContent, HttpError> {
-    let ctx = lock!(ctx.context());
-
-    for e in &ctx.event_channels {
-        e.send(Event::Admin(AdminEvent::Sync)).map_err(|e| {
-            HttpError::for_internal_error(format!("admin event send: {e}"))
-        })?;
-    }
-
-    Ok(HttpResponseUpdatedNoContent())
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema, PartialEq)]
-pub struct EnableStatsRequest {
-    sled_id: Uuid,
-    rack_id: Uuid,
-}
-
-pub const DDM_STATS_PORT: u16 = 8001;
-
-#[endpoint { method = POST, path = "/enable-stats" }]
-async fn enable_stats(
-    ctx: RequestContext<Arc<Mutex<HandlerContext>>>,
-    request: TypedBody<EnableStatsRequest>,
-) -> Result<HttpResponseUpdatedNoContent, HttpError> {
-    let rq = request.into_inner();
-    let ctx = lock!(ctx.context());
-
-    let mut jh = lock!(ctx.stats_handler);
-    if jh.is_none() {
-        let hostname = hostname::get()
-            .expect("failed to get hostname")
-            .to_string_lossy()
-            .to_string();
-        *jh = Some(
-            crate::oxstats::start_server(
-                DDM_STATS_PORT,
-                ctx.peers.clone(),
-                ctx.stats.clone(),
-                hostname,
-                rq.rack_id,
-                rq.sled_id,
-                ctx.log.clone(),
-            )
+        for e in &ctx.event_channels {
+            e.send(Event::Admin(AdminEvent::Announce(PrefixSet::Tunnel(
+                endpoints.clone(),
+            ))))
             .map_err(|e| {
-                HttpError::for_internal_error(format!(
-                    "failed to start stats server: {e}"
-                ))
-            })?,
-        );
+                HttpError::for_internal_error(format!("admin event send: {e}"))
+            })?;
+        }
+
+        match ctx.db.originated_tunnel_count() {
+            Ok(count) => ctx
+                .stats
+                .originated_tunnel_endpoints
+                .store(count as u64, Ordering::Relaxed),
+            Err(e) => {
+                error!(
+                    ctx.log,
+                    "failed to update originated tunnel endpoints stat: {e}"
+                )
+            }
+        }
+        Ok(HttpResponseUpdatedNoContent())
     }
 
-    Ok(HttpResponseUpdatedNoContent())
-}
+    async fn withdraw_prefixes(
+        ctx: RequestContext<Self::Context>,
+        request: TypedBody<HashSet<Ipv6Net>>,
+    ) -> Result<HttpResponseUpdatedNoContent, HttpError> {
+        let ctx = lock!(ctx.context());
+        let prefixes = request.into_inner();
+        ctx.db
+            .withdraw(&prefixes)
+            .map_err(|e| HttpError::for_internal_error(e.to_string()))?;
 
-#[endpoint { method = POST, path = "/disable-stats" }]
-async fn disable_stats(
-    ctx: RequestContext<Arc<Mutex<HandlerContext>>>,
-) -> Result<HttpResponseUpdatedNoContent, HttpError> {
-    let ctx = lock!(ctx.context());
-    let mut jh = lock!(ctx.stats_handler);
-    if let Some(ref h) = *jh {
-        h.abort();
+        for e in &ctx.event_channels {
+            e.send(Event::Admin(AdminEvent::Withdraw(PrefixSet::Underlay(
+                prefixes.clone(),
+            ))))
+            .map_err(|e| {
+                HttpError::for_internal_error(format!("admin event send: {e}"))
+            })?;
+        }
+
+        match ctx.db.originated_count() {
+            Ok(count) => ctx
+                .stats
+                .originated_underlay_prefixes
+                .store(count as u64, Ordering::Relaxed),
+            Err(e) => {
+                error!(
+                    ctx.log,
+                    "failed to update originated underlay prefixes stat: {e}"
+                )
+            }
+        }
+
+        Ok(HttpResponseUpdatedNoContent())
     }
-    *jh = None;
 
-    Ok(HttpResponseUpdatedNoContent())
+    async fn withdraw_tunnel_endpoints(
+        ctx: RequestContext<Self::Context>,
+        request: TypedBody<HashSet<TunnelOrigin>>,
+    ) -> Result<HttpResponseUpdatedNoContent, HttpError> {
+        let ctx = lock!(ctx.context());
+        let endpoints = request.into_inner();
+        slog::info!(ctx.log, "withdraw tunnel: {:#?}", endpoints);
+        ctx.db
+            .withdraw_tunnel(&endpoints)
+            .map_err(|e| HttpError::for_internal_error(e.to_string()))?;
+
+        for e in &ctx.event_channels {
+            e.send(Event::Admin(AdminEvent::Withdraw(PrefixSet::Tunnel(
+                endpoints.clone(),
+            ))))
+            .map_err(|e| {
+                HttpError::for_internal_error(format!("admin event send: {e}"))
+            })?;
+        }
+
+        match ctx.db.originated_tunnel_count() {
+            Ok(count) => ctx
+                .stats
+                .originated_tunnel_endpoints
+                .store(count as u64, Ordering::Relaxed),
+            Err(e) => {
+                error!(
+                    ctx.log,
+                    "failed to update originated tunel endpoints stat: {e}"
+                )
+            }
+        }
+
+        Ok(HttpResponseUpdatedNoContent())
+    }
+
+    async fn sync(
+        ctx: RequestContext<Self::Context>,
+    ) -> Result<HttpResponseUpdatedNoContent, HttpError> {
+        let ctx = lock!(ctx.context());
+
+        for e in &ctx.event_channels {
+            e.send(Event::Admin(AdminEvent::Sync)).map_err(|e| {
+                HttpError::for_internal_error(format!("admin event send: {e}"))
+            })?;
+        }
+
+        Ok(HttpResponseUpdatedNoContent())
+    }
+
+    async fn enable_stats(
+        ctx: RequestContext<Self::Context>,
+        request: TypedBody<EnableStatsRequest>,
+    ) -> Result<HttpResponseUpdatedNoContent, HttpError> {
+        let rq = request.into_inner();
+        let ctx = lock!(ctx.context());
+
+        let mut jh = lock!(ctx.stats_handler);
+        if jh.is_none() {
+            let hostname = hostname::get()
+                .expect("failed to get hostname")
+                .to_string_lossy()
+                .to_string();
+            *jh = Some(
+                crate::oxstats::start_server(
+                    DDM_STATS_PORT,
+                    ctx.peers.clone(),
+                    ctx.stats.clone(),
+                    hostname,
+                    rq.rack_id,
+                    rq.sled_id,
+                    ctx.log.clone(),
+                )
+                .map_err(|e| {
+                    HttpError::for_internal_error(format!(
+                        "failed to start stats server: {e}"
+                    ))
+                })?,
+            );
+        }
+
+        Ok(HttpResponseUpdatedNoContent())
+    }
+
+    async fn disable_stats(
+        ctx: RequestContext<Self::Context>,
+    ) -> Result<HttpResponseUpdatedNoContent, HttpError> {
+        let ctx = lock!(ctx.context());
+        let mut jh = lock!(ctx.stats_handler);
+        if let Some(ref h) = *jh {
+            h.abort();
+        }
+        *jh = None;
+
+        Ok(HttpResponseUpdatedNoContent())
+    }
 }
 
-pub fn api_description() -> Result<
-    ApiDescription<Arc<Mutex<HandlerContext>>>,
-    ApiDescriptionRegisterError,
-> {
-    let mut api = ApiDescription::new();
-    api.register(get_peers)?;
-    api.register(expire_peer)?;
-    api.register(advertise_prefixes)?;
-    api.register(advertise_tunnel_endpoints)?;
-    api.register(withdraw_prefixes)?;
-    api.register(withdraw_tunnel_endpoints)?;
-    api.register(get_prefixes)?;
-    api.register(get_tunnel_endpoints)?;
-    api.register(get_originated)?;
-    api.register(get_originated_tunnel_endpoints)?;
-    api.register(sync)?;
-    api.register(enable_stats)?;
-    api.register(disable_stats)?;
-    Ok(api)
+pub fn api_description(
+) -> Result<ApiDescription<Arc<Mutex<HandlerContext>>>, ApiDescriptionBuildErrors>
+{
+    ddm_admin_api_mod::api_description::<DdmAdminApiImpl>()
 }
