+++ ddm/src/sys.rs
@@ -2,9 +2,9 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-use crate::db::TunnelRoute;
 use crate::sm::{Config, DpdConfig};
 use crate::{dbg, err, inf, wrn};
+use ddm_types::db::TunnelRoute;
 use dpd_client::types;
 use dpd_client::Client;
 use dpd_client::ClientState;
